move a few more to UI
