+++ apps/web/package.json
@@ -40,7 +40,6 @@
     "sharp": "0.33.1",
     "swr": "2.2.4",
     "ui": "workspace:*",
-    "use-breakpoint": "4.0.1",
     "webpack": "5.89.0"
   },
   "devDependencies": {
