+++ apps/web/src/components/Header.tsx
@@ -3,10 +3,10 @@ import { Box, alpha } from '@mui/material';
 import { Section } from 'ui/core/Section';
 import { Nav, NavGroup, NavItem } from 'ui/core/Nav';
 import type { SxProps } from 'ui/theme';
+import { useColorScheme } from 'ui/theme/useColorScheme';
 import { ScrollIndicatorContext } from 'components/ScrollIndicatorContext';
 import { ScrollUpButton } from 'components/ScrollUpButton';
 import { ColorSchemeToggle } from 'components/ColorSchemeToggle';
-import { useColorScheme } from 'hooks/useColorScheme';
 import { Logo } from './Logo';
 
 type HeaderProps = {
