+++ packages/ui/package.json
@@ -9,7 +9,8 @@
   "dependencies": {
     "@mui/material": "5.15.1",
     "react": "18.2.0",
-    "shared-core": "workspace:*"
+    "shared-core": "workspace:*",
+    "use-breakpoint": "4.0.1"
   },
   "devDependencies": {
     "@types/react": "18.2.45",
